==> data/Forecast/Step 1 Impute Data.ipynb <==
import pandas as pd
import numpy as np
import sqlite3
from itertools import product
pd.set_option('display.max_columns', 500)

import matplotlib.pyplot as plt

conn = sqlite3.connect(r'C:\Users\Spencer\Environments\aquastaat\project.db')
conn.text_factory = lambda x: str(x, 'latin1')

data = pd.read_sql("SELECT * FROM parameter", conn)

def remove_blanks(data):
    print('Before: ', data.shape)
    data.loc[data.loc[:,'Value']=='','Value'] = None
    data.dropna(axis=0, how='any', inplace=True)
    print('After: ', data.shape)
    
def label_bin(year, year_bins, bin_labels):
    return bin_labels[len(list(filter(lambda y: y < year, year_bins)))-1]
    
def bin_year(data):
    year_bins = [y for y in range(1978, 2020, 5)]
    bin_labels = [l for l in range(1980, 2020, 5)]
    data['YearBin'] = data['Year'].apply(lambda x: label_bin(x, year_bins, bin_labels))

data = data.loc[data['Year']>=1978,:]

remove_blanks(data)
bin_year(data)
data['Value'] = data['Value'].astype(float)

data.drop(axis=1, labels=['VariableId', 'AreaId', 'Year'], inplace=True)

All = slice(None)
WSI = 'SDG 6.4.2. Water Stress'
TFW = 'Total freshwater withdrawal (primary and secondary)'

tfw_predictors = [
    'Total population',
    'Cultivated area (arable land + permanent crops)',
    'Gross Domestic Product (GDP)',
    'Total renewable surface water',
]

data = data.pivot_table(index=['Area', 'YearBin'], columns='VariableName', values='Value', aggfunc=np.mean)

data = data.loc[:, tfw_predictors + [WSI, TFW]]

data

areas = data.index.levels[0]
bin_labels = [l for l in range(1980, 2020, 5)]
training_data = pd.DataFrame(list(product(areas, bin_labels)), columns=['Area', 'YearBin'])
training_data.set_index(['Area', 'YearBin'], inplace=True)
training_data = training_data.join(data, how='left')
training_data = training_data.groupby('Area').apply(lambda x: x.interpolate(limit_direction='both'))
training_data.dropna(axis=0, inplace=True)

training_data.to_csv('training_data.csv')

training_data
