==> data/Forecast/Step 2 Countries.ipynb <==
import pandas as pd
import sqlite3
import numpy as np
from itertools import product
pd.set_option("display.max_columns", 500)

data = pd.read_csv('training_data.csv')

countries = data.loc[:, 'Area'].unique()

gdp = pd.read_excel('countries_gdp.xlsx')
pop = pd.read_excel('population_growth.xlsx')

gdp.sort_values(by='country', inplace=True)
pop.sort_values(by='COUNTRY', inplace=True)

if input("load maps? y/n") == 'y':
    gdp = pd.read_csv('gdp_mapped.csv')
    pop = pd.read_csv('pop_mapped.csv')

if input("create new maps? y/n") == 'y':
    gdp_map = {}
    for c in gdp.itertuples():
        if c.country in countries:
            gdp_map[c.country] = c.country
        else:
            while gdp_map.get(c.country) not in countries and gdp_map.get(c.country) != 'nan':
                gdp_map[c.country] = input(c.country)    
    gdp['mapped'] = gdp['country'].apply(lambda x: gdp_map.get(x, 'nan'))
    gdp = gdp.loc[gdp.loc[:,'mapped']!='nan',:]
    
    pop_map = {}
    for c in pop.itertuples():
        if c.COUNTRY in countries:
            pop_map[c.COUNTRY] = c.COUNTRY
        else:
            while pop_map.get(c.COUNTRY) not in countries and pop_map.get(c.COUNTRY) != 'nan':
                pop_map[c.COUNTRY] = input(c.COUNTRY)
    pop['mapped'] = pop['COUNTRY'].apply(lambda x: pop_map.get(x, 'nan'))
    pop = pop.loc[pop.loc[:,'mapped']!='nan',:]          

if input("save maps again? y/n") == 'y':
    gdp.to_csv('gdp_mapped.csv', index=False)
    pop.to_csv('pop_mapped.csv', index=False)

pop.shape, gdp.shape, len(countries)

pop.set_index('mapped', inplace=True)
gdp.set_index('mapped', inplace=True)

pop['POPULATION'] = pop['POPULATION'].apply(lambda x: float("-" + x[1:]) if "−" in str(x) else x)

years = np.concatenate([data['YearBin'].unique(),[2020, 2025, 2030]])

def forecast(start, rate, years):
    return start*(1+rate/100)**years

new_data = pd.DataFrame(list(product(countries, years)), columns = ['Area', 'YearBin'])
new_data.set_index(['Area', 'YearBin'], inplace=True)
new_data = new_data.join(data.set_index(['Area', 'YearBin']), how='left')

All = slice(None)
col_pop = 'Total population'
for name, group in new_data.groupby('Area'):
    last_val = group.at[(name, 2015), col_pop]
    rate = pop.at[name, 'POPULATION']
    new_data.at[(name, 2020), col_pop] = forecast(last_val, rate, 5)
    new_data.at[(name, 2025), col_pop] = forecast(last_val, rate, 10)
    new_data.at[(name, 2030), col_pop] = forecast(last_val, rate, 15)

col_gdp = 'Gross Domestic Product (GDP)'
world_gdp = 3.2
for name, group in new_data.groupby('Area'):
    last_val = group.at[(name, 2015), col_gdp]
    if name in gdp.index:
        rate = gdp.at[name, 'gdp_rate']
    else:
        print(name)
        rate = world_gdp
    new_data.at[(name, 2020), col_gdp] = forecast(last_val, rate, 5)
    new_data.at[(name, 2025), col_gdp] = forecast(last_val, rate, 10)
    new_data.at[(name, 2030), col_gdp] = forecast(last_val, rate, 15)

new_data

new_data.to_csv('training_data_rate.csv')
