==> data/Forecast/Step 3 ARIMA.ipynb <==
import pandas as pd
import numpy as np
pd.set_option("display.max_columns", 500)

import matplotlib.pyplot as plt

data = pd.read_csv('training_data_rate.csv')
data.set_index(['Area', 'YearBin'], inplace=True)

All = slice(None)
c_area = 'Cultivated area (arable land + permanent crops)'
cult = pd.read_csv('predicted_cultivated_area_v2.csv', encoding='latin1')
cult = cult.iloc[:, 1:]
cult.set_index('Area', inplace=True)

for name, row in cult.iterrows():
    data.at[(name, 2020), c_area] = row['Y2020']
    data.at[(name, 2025), c_area] = row['Y2025']
    data.at[(name, 2030), c_area] = row['Y2030']

c_trsw = 'Total renewable surface water'
trsw = pd.read_csv('predicted_TRSW_v2.csv', encoding='latin1')
trsw = trsw.iloc[:, 1:]
trsw.set_index('Area', inplace=True)

for name, row in trsw.iterrows():
    data.at[(name, 2020), c_trsw] = row['Y2020']
    data.at[(name, 2025), c_trsw] = row['Y2025']
    data.at[(name, 2030), c_trsw] = row['Y2030']

data = data.loc[(~data.loc[:,'Total population'].isna()) &
                (~data.loc[:, c_area].isna()),:]

data.to_csv('training_data_arima.csv')
